requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-rq==2.10.2
rq==1.16.2
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
